requirements.txt
accelerate>=0.29
bitsandbytes>=0.43
datasets>=2.19
mauve-text>=0.3
numpy>=1.24
openai>=1.30
pandas>=2.0
peft>=0.10
python-dotenv>=1.0
pyyaml>=6.0
rapidfuzz>=3.9
sacremoses>=0.0.53
scikit-learn>=1.4
sentence-transformers>=2.7
sentencepiece>=0.1.99
tenacity>=8.2
tqdm>=4.66
transformers>=4.38
trl>=0.8
